+++ core/API/DataTableManipulator/Flattener.php
@@ -38,17 +38,16 @@ class Flattener extends DataTableManipulator
      * Separator for building recursive labels (or paths)
      * @var string
      */
-    public $recursiveLabelSeparator = ' - ';
+    public $recursiveLabelSeparator = '';
 
     /**
      * @param  DataTable $dataTable
+     * @param string $recursiveLabelSeparator
      * @return DataTable|DataTable\Map
      */
-    public function flatten($dataTable)
+    public function flatten($dataTable, $recursiveLabelSeparator)
     {
-        if ($this->apiModule == 'Actions' || $this->apiMethod == 'getWebsites') {
-            $this->recursiveLabelSeparator = '/';
-        }
+        $this->recursiveLabelSeparator = $recursiveLabelSeparator;
 
         return $this->manipulate($dataTable);
     }
@@ -72,9 +71,10 @@ class Flattener extends DataTableManipulator
         }
 
         $newDataTable = $dataTable->getEmptyClone($keepFilters);
-        foreach ($dataTable->getRows() as $row) {
-            $this->flattenRow($row, $newDataTable);
+        foreach ($dataTable->getRows() as $rowId => $row) {
+            $this->flattenRow($row, $rowId, $newDataTable);
         }
+
         return $newDataTable;
     }
 
@@ -84,15 +84,21 @@ class Flattener extends DataTableManipulator
      * @param string $labelPrefix
      * @param bool $parentLogo
      */
-    private function flattenRow(Row $row, DataTable $dataTable,
+    private function flattenRow(Row $row, $rowId, DataTable $dataTable,
                                 $labelPrefix = '', $parentLogo = false)
     {
         $label = $row->getColumn('label');
         if ($label !== false) {
             $label = trim($label);
-            if (substr($label, 0, 1) == '/' && $this->recursiveLabelSeparator == '/') {
-                $label = substr($label, 1);
+
+            if ($this->recursiveLabelSeparator == '/') {
+                if (substr($label, 0, 1) == '/') {
+                    $label = substr($label, 1);
+                } elseif ($rowId === DataTable::ID_SUMMARY_ROW && $labelPrefix && $label != DataTable::LABEL_SUMMARY_ROW) {
+                    $label = ' - ' . $label;
+                }
             }
+
             $label = $labelPrefix . $label;
             $row->setColumn('label', $label);
         }
@@ -103,7 +109,16 @@ class Flattener extends DataTableManipulator
             $row->setMetadata('logo', $logo);
         }
 
-        $subTable = $this->loadSubtable($dataTable, $row);
+        /** @var DataTable $subTable */
+        $subTable = $row->getSubtable();
+
+        if ($subTable) {
+            $subTable->applyQueuedFilters();
+            $row->deleteMetadata('idsubdatatable_in_db');
+        } else {
+            $subTable = $this->loadSubtable($dataTable, $row);
+        }
+
         $row->removeSubtable();
 
         if ($subTable === null) {
@@ -117,8 +132,8 @@ class Flattener extends DataTableManipulator
                 $dataTable->addRow($row);
             }
             $prefix = $label . $this->recursiveLabelSeparator;
-            foreach ($subTable->getRows() as $row) {
-                $this->flattenRow($row, $dataTable, $prefix, $logo);
+            foreach ($subTable->getRows() as $rowId => $row) {
+                $this->flattenRow($row, $rowId, $dataTable, $prefix, $logo);
             }
         }
     }